+++ include/xlink2/xlink2SystemELink.h
@@ -35,7 +35,7 @@ public:
     void allocHandle(sead::Heap*) override;
     void allocAssetExecutor(Event*) override;
 
-    u64 getResourceVersion() const override;
+    u32 getResourceVersion() const override;
     sead::SafeString* getModuleName() const override;
 
 #ifdef SEAD_DEBUG
