+++ src/xlink2SystemELink.cpp
@@ -1,7 +1,7 @@
 #include "xlink2/xlink2SystemELink.h"
 
 namespace xlink2 {
-u64 SystemELink::getResourceVersion() const {
+u32 SystemELink::getResourceVersion() const {
     return 0x1e;
 }
